Move log stream to dedicated pkg
